refactor(things test): seperating things routes into its own test spec
