+++ server/api/feelings.spec.js
@@ -1,7 +1,7 @@
 const {expect} = require('chai')
 const request = require('supertest')
 const db = require('../db')
-const app = require('../index')
+const app = require('../')
 const Feeling = db.model('feeling')
 
 describe('Feeling routes', () => {
